Добавляю преобразование типа модуля в макрос P7_LOG для более удобного его использования

+++ Singleton/p7_logger.h
@@ -77,7 +77,7 @@ private:
 #ifdef USE_P7_LOG
 #define P7_LOG(log_level, module_enum, format, ...)\
    p7_logger_raii::instance().trace().P7_DELIVER(0,\
-                                            log_level,\
+                                            static_cast<size_t>(log_level),\
                                             p7_logger_raii::instance().module(module_enum),\
                                             format,\
                                             ##__VA_ARGS__)
